requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-environ==0.4.4
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
mysqlclient==1.3.12
oauthlib==2.0.6
PyJWT==1.5.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
